Move packages to spark.index.query and add test for overwrite
