+++ src/main/scala/io/qbeast/spark/index/query/QueryIndexStatusExecutor.scala
@@ -1,10 +1,10 @@
 /*
  * Copyright 2021 Qbeast Analytics, S.L.
  */
-package io.qbeast.spark.internal
+package io.qbeast.spark.index.query
 
 import io.qbeast.IISeq
-import io.qbeast.core.model.{CubeId, CubeStatus, IndexStatus, QbeastFile, QuerySpace, WeightRange}
+import io.qbeast.core.model._
 import io.qbeast.spark.utils.State
 
 /**
