+++ src/main/scala/io/qbeast/spark/index/query/QuerySpec.scala
@@ -1,12 +1,13 @@
 /*
  * Copyright 2021 Qbeast Analytics, S.L.
  */
-package io.qbeast.spark.internal
+package io.qbeast.spark.index.query
 
 import io.qbeast.core.model.{QuerySpace, WeightRange}
 
 /**
  * Query specification
+ *
  * @param weightRange the weight range
  * @param querySpace the query space
  */
